+++ nft-contract-royalty/src/metadata.rs
@@ -44,9 +44,9 @@ pub struct Token {
     //owner of the token
     pub owner_id: AccountId,
     //list of approved account IDs that have access to transfer the token. This maps an account ID to an approval ID
-    pub approved_account_ids: HashMap<AccountId, u64>,
+    pub approved_account_ids: HashMap<AccountId, u32>,
     //the next approval ID to give out. 
-    pub next_approval_id: u64,
+    pub next_approval_id: u32,
     //keep track of the royalty percentages for the token in a hash map
     pub royalty: HashMap<AccountId, u32>,
 }
@@ -62,7 +62,7 @@ pub struct JsonToken {
     //token metadata
     pub metadata: TokenMetadata,
     //list of approved account IDs that have access to transfer the token. This maps an account ID to an approval ID
-    pub approved_account_ids: HashMap<AccountId, u64>,
+    pub approved_account_ids: HashMap<AccountId, u32>,
     //keep track of the royalty percentages for the token in a hash map
     pub royalty: HashMap<AccountId, u32>,
 }
